LPS-73853 Move to a better package to match com.liferay.portal.dao.orm.hibernate.DB2Dialect
